# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ELOJIM API
  description: API oficial del proyecto Fundación ELOJIM
  version: "1.0.0"
  contact:
    name: Soporte ELOJIM
    email: soporte@elojim.org
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    RegistroBase:
      type: object
      properties:
        nombreCompleto:
          type: string
          example: "Laura Gómez"
        fechaNacimiento:
          type: string
          format: date
        edad:
          type: integer
        comuna:
          type: string
        estratoSocial:
          type: string
        grupoEtnico:
          type: string
        telefonoContacto:
          type: string
        correoElectronico:
          type: string
        direccion:
          type: string
        motivacion:
          type: string
        expectativas:
          type: string
        aceptaTerminos:
          type: boolean
      required:
        - nombreCompleto
        - fechaNacimiento
        - comuna
        - estratoSocial
        - telefonoContacto
        - correoElectronico
        - aceptaTerminos
...
    RegistroEconomiaPlateada:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            tipoDocumento:
              type: string
            numeroDocumento:
              type: string
            genero:
              type: string
            esPensionado:
              type: boolean
            actividadEconomica:
              type: string
            trabajoAnterior:
              type: boolean
            sectorTrabajo:
              type: string
            ingresosAdicionales:
              type: boolean
            fuenteIngresos:
              type: string
            areasInteres:
              type: array
              items:
                type: string
            otrasAreas:
              type: string
            habilidades:
              type: string
            tiempoSemanal:
              type: string
          example:
            nombreCompleto: "Sandra Perdomo"
            esPensionado: true
            actividadEconomica: "Venta de productos naturales"
            habilidades: "Manualidades y cocina"
            aceptaTerminos: true
...
security:
  - cookieAuth: []
paths:
...
  /api/registro/refuerzo-escolar:
    get:
      tags: [Registro]
      summary: Listar registros de refuerzo escolar
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro en refuerzo escolar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreCompleto: { type: string, example: "Ernesto Pérez" }
                fechaNacimiento: { type: string, format: date-time, example: "2011-05-15T00:00:00Z" }
                comuna: { type: string, example: "Comuna 10" }
                estratoSocial: { type: string, example: "E3" }
                edad: { type: integer, example: 15 }
                grupoEtnico: { type: string, example: "Rom_Gitano" }
                institucionEducativa: { type: string, example: "Colegio ABC" }
                cursoGrado: { type: string, example: "9" }
                direccion: { type: string, example: "Calle 123, Barrio XYZ" }
                nombreAcudiente: { type: string, example: "Ana Pérez" }
                relacionNino: { type: string, example: "Madre" }
                telefonoContacto: { type: string, example: "3012345671" }
                correoElectronico: { type: string, example: "ana.perez@example.com" }
                areasAyuda:
                  type: array
                  items: { type: string }
                  example: ["Matemáticas", "Lectura y comprensión", "Inglés"]
                otrasAreas: { type: string, example: "Tecnologías de la información" }
                refuerzoPrevio: { type: boolean, example: true }
                dificultadesAcademicas: { type: string, example: "Problemas con matemáticas y lectura" }
                disponibilidad: { type: string, example: "Lunes a Viernes, 3:00 PM - 5:00 PM" }
                accesoMateriales: { type: boolean, example: true }
                apoyoHabitos: { type: boolean, example: true }
                motivacion: { type: string, example: "Me gustaría mejorar mis notas y aprender más en las materias" }
                expectativas: { type: string, example: "Espero mejorar mis calificaciones y aprender de forma más divertida" }
                aceptaTerminos: { type: boolean, example: true }
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "401":
          description: No autenticado
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de refuerzo escolar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno
...
  /api/registro/economia-plateada:
    get:
      tags: [Registro]
      summary: Listar registros del programa economía plateada
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro para Economia Plateada
      description: Crear nuevo registro para el programa Economia Plateada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroEconomiaPlateada"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de programa economia plateada
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno      